pyproject.toml
[project]
name = "e2e-fraud-detection"
version = "1.0.0"
description = "End to End credit card fraud detection system powered by Azure ML"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [{ name = "Temuulen Tuvshinjargal", email = "codetemuulen@gmail.com" }]

dependencies = [
    "azure-cli==2.81.0",
    "azure-ai-ml==1.30.0",
    "azure-identity==1.17.0",
    "mlflow==3.7.0",
    "pandas==2.3.3",
    "pynacl==1.5.0",
    "requests==2.32.5",
    "scikit-learn==1.8.0",
    "lightgbm==4.6.0",
    "xgboost==3.1.2",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "typer==0.20.0",
    "mltable==1.6.3",
    "azureml-mlflow==1.60.0",
    "pyyaml==6.0.2",
]

[project.optional-dependencies]
dev = [
    "black==24.10.0",
    "pip-audit==2.7.0",
    "ruff==0.7.0",
    "pytest==8.3.0",
    "pytest-mock==3.14.0",
    "ipykernel==7.1.0",
    "seaborn==0.13.2",
]

[project.urls]
Repository = "https://github.com/CodeTemka/e2e-fraud-detection"

[project.scripts]
fraud-cli = "fraud_detection.cli:run"


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["fraud_detection*"]

[build-system]
requires = ["setuptools==80", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["data", "notebook"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "cli: marks command-line interface tests with mocked azure calls",
    "integration: marks tests that require Azure ML env vars or live resources",
]
